requirements.txt
asgiref==3.5.2
astroid==2.9.3

autopep8==2.0.0
black==22.8.0
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-js-asset==2.0.0
django-stubs==1.12.0
django-stubs-ext==0.5.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.1
drf-yasg==1.21.4
greenlet==1.1.3
idna==3.4
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
msgpack==1.0.4
mypy==0.982
mypy-extensions==0.4.3
mysqlclient==2.1.1
oauthlib==3.2.1
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.1
ply==3.11
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.5.0
pylint==2.12.2
pynvim==0.4.3
PyQt5==5.15.7
python3-openid==3.2.0
pytz==2022.4
pywin32==302
requests==2.28.1
requests-oauthlib==1.3.1
rest-framework-roles==1.0.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
types-pytz==2022.2.1.0
types-PyYAML==6.0.12
typing_extensions==4.3.0
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
wincertstore==0.2
wrapt==1.13.3